requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.8.0
python-telegram-bot==20.4